Fix docs failing in GA
In this commit we update the packages of sphinx
and sphinx-rtd-theme to resolve a package
dependency issue that caused the building of the
docs to fail.
The docs/requirements.txt file was generated by
creating a virtual env and installing only
sphinx and sphinx-rtd-theme and the running
pipenv lock -r to get the requirements and copy
paste them into the docs/requirements.txt.

Fixes #73